requirements.txt
mypy
pylint
python-dotenv
typer
